backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0